requirements.txt
asgiref==3.6.0
attrs==22.2.0
autopep8==2.0.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.7
djangorestframework==3.14.0
drf-yasg==1.21.5
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.2
mysqlclient==2.1.1
openapi-codec==1.3.2
packaging==23.0
pycodestyle==2.10.0
pydotplus==2.0.2
pyparsing==3.0.9
pyrsistent==0.19.3
pytz==2022.7.1
PyYAML==6.0
requests==2.15.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
simplejson==3.18.3
sqlparse==0.4.3
swagger-spec-validator==3.0.3
tomli==2.0.1
typing_extensions==4.5.0
tzdata==2022.7
uritemplate==4.1.1
gunicorn
pillow
